@@ -50,10 +50,11 @@
     "condition-circle": "^1.5.0",
     "cross-env": "^5.1.1",
     "cz-conventional-changelog": "^2.0.0",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^7.1.0",
+    "eslint": "^4.10.0",
+    "eslint-config-standard": "^10.2.1",
     "eslint-plugin-babel": "^4.0.0",
     "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-standard": "^3.0.1",
     "metalsmith-markdown": "^0.2.1",
